move display metod inside displayer module
